backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
mysql-connector-python
requests
python-dotenv
marshmallow
Flask-Migrate
python-dotenv
bcrypt
flask_cors
